Added sorting algorithm that accounts for child entries' scores in computing the parent entry's score
Added ability to count children

Added round() to round floats to an integer number of digits

Updated DescendantEntriesChildParent query to be able to limit
the number of returned entries. The results will pull down N
children of the targeted entry, and then all children of those
N children. This is useful for saying "Give me the most recent
20 records, and all of their children." Obviously this is not
infinitely scaleable but this should be plenty for now.